Group comparison tab in results view
