+++ src/pages/groupComparison/MutationEnrichments.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
 import { observer } from 'mobx-react';
-import GroupComparisonStore from './GroupComparisonStore';
 import { MolecularProfile } from '../../shared/api/generated/CBioPortalAPI';
 import LoadingIndicator from '../../shared/components/loadingIndicator/LoadingIndicator';
 import ErrorMessage from '../../shared/components/ErrorMessage';
@@ -12,9 +11,12 @@ import { MakeEnrichmentsTabUI } from './GroupComparisonUtils';
 import { remoteData } from 'cbioportal-frontend-commons';
 import _ from 'lodash';
 import { AlterationContainerType } from 'pages/resultsView/enrichments/EnrichmentsUtil';
+import ComparisonStore from '../../shared/lib/comparison/ComparisonStore';
+import { ResultsViewPageStore } from '../resultsView/ResultsViewPageStore';
 
 export interface IMutationEnrichmentsProps {
-    store: GroupComparisonStore;
+    store: ComparisonStore;
+    resultsViewStore?: ResultsViewPageStore;
 }
 
 @observer
@@ -104,6 +106,7 @@ export default class MutationEnrichments extends React.Component<
                         onSetPatientLevelEnrichments={
                             this.props.store.setUsePatientLevelEnrichments
                         }
+                        store={this.props.resultsViewStore}
                     />
                 </div>
             );
