+++ src/pages/resultsView/comparison/ResultsViewComparisonUtils.ts
@@ -1,0 +1,6 @@
+import { ComparisonGroup } from '../../groupComparison/GroupComparisonUtils';
+
+export type ResultsViewComparisonGroup = ComparisonGroup & {
+    nameOfEnrichmentDirection: string;
+    count: number;
+};
