+++ src/pages/resultsView/enrichments/AddCheckedGenes.tsx
@@ -18,10 +18,10 @@ export default class AddCheckedGenes extends React.Component<
     @autobind
     private onAddGenes() {
         // add genes and go back to oncoprint tab
-        (window as any).routingStore.updateRoute(
+        (window as any).urlWrapper.updateURL(
             {
                 [ResultsViewURLQueryEnum.gene_list]: `${
-                    (window as any).routingStore.query[
+                    (window as any).urlWrapper.query[
                         ResultsViewURLQueryEnum.gene_list
                     ]
                 }\n${this.props.checkedGenes.join(' ')}`,
